apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.2
Pillow==10.4.0

# YOLOv8
ultralytics==8.3.0

# LangChain + LangGraph
langchain-core==0.3.0
langchain==0.3.0
langgraph==0.2.28
langchain-google-genai==2.0.0
langchain-openai==0.2.0

# ZhipuAI
zhipuai==2.1.5

# Database (for estimation tools querying scraped prices)
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Image processing
opencv-python-headless==4.10.0.84
numpy==1.26.4
